// microfrontends/api-middleware/src/Chat/action-types.ts
export enum ChatActionTypeEnum {
  API_CHAT_CONNECTION = "API/CHAT/CONNECTION",
  CHAT_SET_CONNECTION_MESSAGE = "CHAT_SET_CONNECTION_MESSAGE",

  API_CHAT_GET_HISTORY_DIALOG = "API/CHAT/GET_HISTORY_DIALOG",
  CHAT_SET_HISTORY_DIALOG = "CHAT_SET_HISTORY_DIALOG",

  API_CHAT_NEW_RESPONSE_MESSAGE = "API/CHAT/NEW_RESPONSE_MESSAGE",

  API_CHAT_NEW_REQUEST_MESSAGE = "API/CHAT/NEW_REQUEST_MESSAGE",

  API_CHAT_SEND_FILE = "API/CHAT/SEND_FILE",
  CHAT_SEND_FILE_STATUS = "CHAT_SEND_FILE_STATUS",

  CHAT_SET_ERROR = "CHAT_SET_ERROR",

  API_CHAT_REFRESH_ERROR = "API/CHAT_REFRESH_ERROR",
  API_CHAT_REFRESH_FILE_STATUS = "API/CHAT_REFRESH_FILE_STATUS",

  API_CHAT_CLIENT_ACTIVITY = "API/CHAT/CLIENT_ACTIVITY",

  CHAT_UNREAD_MESSAGES_COUNT = "CHAT_UNREAD_MESSAGES_COUNT",

  API_CHAT_SET_READ_MESSAGES = "API/CHAT/SET_READ_MESSAGES",

  API_CHAT_ACTUAL_APP_PATH = "API/CHAT/ACTUAL_APP_PATH",
  CHAT_SET_ACTUAL_APP_PATH = "CHAT_SET_ACTUAL_APP_PATH",
}