// microfrontends/api-middleware/src/Login/action-types.ts
export enum LoginActionType {
  API_LOGIN_CHECK_AUTH = "API/LOGIN/CHECK_AUTH",

  API_LOGIN_SIGN_IN = "API/LOGIN/SIGN_IN",
  SIGN_IN_START = "SIGN_IN_START",
  SIGN_IN_ACCEPTED = "SIGN_IN_ACCEPTED",
  SIGN_IN_DENIED = "SIGN_IN_DENIED",

  API_LOGIN_SIGN_OUT = "API/LOGIN/SIGN_OUT",

  API_LOGIN_FORGOT_PASSWORD = "API/LOGIN/FORGOT_PASSWORD",

  API_LOGIN_RECREATE_PASSWORD = "API/LOGIN/RECREATE_PASSWORD",
  RECREATE_PASSWORD_START = "RECREATE_PASSWORD_START",
  RECREATE_PASSWORD_ACCEPTED = "RECREATE_PASSWORD_ACCEPTED",
  RECREATE_PASSWORD_DENIED = "RECREATE_PASSWORD_DENIED",

  API_LOGIN_REFRESH_STATE = "API/LOGIN/REFRESH_STATE",
}